requirements.txt
aiohttp==3.11.12
beautifulsoup4==4.13.3
discord.py==2.4.0
psutil==6.1.1
python-dotenv==1.0.1
requests==2.32.3
python-telegram-bot==21.10